polish

+++ project/bundles/appimage/01-build-centos6.sh
@@ -41,10 +41,13 @@ if [[ "$(arch)" = "x86_64" ]] ; then
     yum upgrade ca-certificates --disablerepo=epel
 fi
 
-yum install epel-release
+if [[ ! -f /etc/yum.repos.d/epel.repo ]] ; then
 
-# we need to be up to date in order to install the xcb-keysyms dependency
-yum -y update
+    yum install epel-release
+
+    # we need to be up to date in order to install the xcb-keysyms dependency
+    yum -y update
+fi
 
 #################################################################################################
